// src/api/generated/models/calendarEventType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type CalendarEventType = typeof CalendarEventType[keyof typeof CalendarEventType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CalendarEventType = {
  acquisition: 'acquisition',
  appointment: 'appointment',
  generic_appointment: 'generic_appointment',
  sale_preliminary: 'sale_preliminary',
  property_viewing: 'property_viewing',
  call_scheduled: 'call_scheduled',
  call_completed: 'call_completed',
  rental_proposal: 'rental_proposal',
  purchase_proposal: 'purchase_proposal',
} as const;