// src/api/generated/models/calendarEventConfirmationStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

/**
 * @nullable
 */
export type CalendarEventConfirmationStatus = typeof CalendarEventConfirmationStatus[keyof typeof CalendarEventConfirmationStatus] | null;


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CalendarEventConfirmationStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;