// src/api/generated/models/calendarEventReminder.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type CalendarEventReminder = typeof CalendarEventReminder[keyof typeof CalendarEventReminder];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CalendarEventReminder = {
  none: 'none',
  at_time: 'at_time',
  '5_minutes_before': '5_minutes_before',
  '15_minutes_before': '15_minutes_before',
  '30_minutes_before': '30_minutes_before',
  '1_hour_before': '1_hour_before',
  '1_day_before': '1_day_before',
} as const;